app/requirements.txt
Django==3.2.13
django-webpack-loader==3.1.0
djangorestframework~=3.12.2
psycopg2-binary==2.9.3
django-cors-headers==3.10.0
gunicorn==20.1.0
coreapi==2.3.3